refactor: フォーマッタ用の整形関数をヘルパー関数lからの読み込みに修正
- 複数のフォーマッターファイルでAgToLabelのインポートパスを修正
- MockFormatterのJSON出力におけるlogLevelプロパティの配置を並べ替えおよび調整
- PlainFormatterでローカル実装されていたargsToStringを共通ユーティリティのインポートに置き換え
- AgLogValidatorsのvalueToString関数のインポートを共通ヘルパーに移動し、ローカル定義を削除
